phase-2-web/backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
email-validator==2.1.0

# CORS
python-multipart==0.0.6

# Development
pytest==7.4.4
httpx==0.26.0
